api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.16",
    "apscheduler>=3.11.0",
    "bcrypt>=4.3.0",
    "dnspython>=2.7.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "gitpython>=3.1.44",
    "jose>=1.0.0",
    "motor>=3.7.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
